Updated variable types to get more accurate answer in system out message

+++ src/edu/dmacc/coma510/Problem1.java
@@ -6,7 +6,7 @@ public class Problem1 {
         int numberOfWalks = 4;
         int numberOfHits = 24;
 
-        int calculatedBattingAverage = numberOfHits / (numberOfPlateAppearances - numberOfWalks);
+        double calculatedBattingAverage = (double)(numberOfHits) / (numberOfPlateAppearances - numberOfWalks);
 
         System.out.println(calculatedBattingAverage); // 0.25
     }
